+++ saviSimulation/savi/StateSynchronization/TimePerception.java
@@ -3,7 +3,7 @@ package savi.StateSynchronization;
 
 public class TimePerception extends Perception {
 
-	public TimePerception(long timeStamp) {
+	public TimePerception(double timeStamp) {
 		super("time", null, timeStamp, null);
 	}
 }
